backend-py/requirements.txt
langchain==0.1.5
langchain-community==0.0.19
langchain-openai==0.0.3
uvicorn==0.23.2
beautifulsoup4==4.12.2
weaviate-client==3.23.2
psycopg2==2.9.7
lxml==4.9.3
python-dotenv==1.0.1
fake_useragent==1.4.0
